refactor(core,agent): move loop and stagnation detectors into gglib-core domain
The loop detector (FNV-1a batch signatures over canonicalized tool-call
arguments, dual observation/action thresholds) and the stagnation detector
(session-wide response-hash occurrence counts, catches A->B->A->B
oscillation) have lived as pub(crate) modules inside gglib-agent, guarding
only the built-in agent paths (gglib chat, gglib q, the web chat). The
proxy path has no access to them, so external agentic clients get no loop
protection at all.

This moves fnv1a, loop_detection/ and stagnation/ verbatim into
gglib_core::domain::agent, where their only dependencies (ToolCall,
AgentError, serde_json) already live -- zero new crate edges. check() and
record() widen from pub(crate) to pub; signatures are unchanged, so the
23 unit tests move verbatim and gglib-agent's Guards and the
integration_guards suite are untouched semantically. gglib-agent now
imports the detectors from core.

Placing them in core follows the request-pipeline precedent: one
implementation shared by the agent path and (next) the proxy path, so the
two cannot drift. The moved code returns AgentError from crate::ports --
an intra-crate domain->ports import, legal and cycle-free; the
alternative (detector-local verdict enums mapped by callers) was rejected
to keep the move behavior-preserving.

No behavior change.

Parity: none (pure move)
