+++ crates/gglib-agent/src/lib.rs
@@ -7,9 +7,6 @@
 
 pub(crate) mod agent_loop;
 pub(crate) mod context_pruning;
-pub(crate) mod fnv1a;
-pub(crate) mod loop_detection;
-pub(crate) mod stagnation;
 pub(crate) mod stream_collector;
 pub(crate) mod tool_execution;
 pub(crate) mod util;
