+++ crates/gglib-core/src/domain/agent/fnv1a.rs
@@ -1,13 +1,10 @@
-//! FNV-1a 64-bit hash utility used across the `gglib-agent` crate.
+//! FNV-1a 64-bit hash utility backing the loop and stagnation detectors.
 //!
 //! # Why inline rather than using the `fnv` crate?
 //!
 //! The full implementation is 10 lines with no unsafe code and no trait
-//! machinery.  Vendoring it keeps the `gglib-agent` dependency graph minimal
+//! machinery.  Vendoring it keeps the `gglib-core` dependency graph minimal
 //! and avoids pulling in an additional crate for a single trivial function.
-//!
-//! Timing utilities live in `gglib_core::utils::timing` and are re-exported
-//! from the `gglib_core` crate root as `gglib_core::elapsed_ms`.
 
 // =============================================================================
 // FNV-1a 64-bit hash
