+++ crates/gglib-core/src/domain/agent/loop_detection/tests.rs
@@ -3,7 +3,7 @@ use serde_json::json;
 use super::*;
 
 // ---- tool_signature / batch_signature ---------------------------------------
-// Note: fnv1a_64 correctness is covered in crates/gglib-agent/src/fnv1a.rs.
+// Note: fnv1a_64 correctness is covered in crates/gglib-core/src/domain/agent/fnv1a.rs.
 
 #[test]
 fn tool_signature_includes_name_and_hash() {
