+++ crates/gglib-core/src/domain/agent/mod.rs
@@ -1,8 +1,11 @@
 #![doc = include_str!("README.md")]
 pub mod config;
 pub mod events;
+pub mod fnv1a;
+pub mod loop_detection;
 pub mod messages;
 mod messages_serde;
+pub mod stagnation;
 pub mod tool_display;
 pub mod tool_types;
 
@@ -13,5 +16,7 @@ pub use config::{
     MAX_TOOL_TIMEOUT_MS_CEILING, MIN_CONTEXT_BUDGET_CHARS, MIN_TOOL_TIMEOUT_MS,
 };
 pub use events::{AGENT_EVENT_CHANNEL_CAPACITY, AgentEvent, LlmStreamEvent};
+pub use loop_detection::{LoopDetector, is_observation_batch};
 pub use messages::{AgentMessage, AssistantContent};
+pub use stagnation::StagnationDetector;
 pub use tool_types::{ToolCall, ToolDefinition, ToolResult};
