+++ crates/gglib-core/src/domain/agent/stagnation/README.md
@@ -1,7 +1,7 @@
 # stagnation
 
-![LOC](https://img.shields.io/endpoint?url=https://raw.githubusercontent.com/mmogr/gglib/badges/gglib-agent-stagnation-loc.json)
-![Complexity](https://img.shields.io/endpoint?url=https://raw.githubusercontent.com/mmogr/gglib/badges/gglib-agent-stagnation-complexity.json)
+![LOC](https://img.shields.io/endpoint?url=https://raw.githubusercontent.com/mmogr/gglib/badges/gglib-core-domain-agent-stagnation-loc.json)
+![Complexity](https://img.shields.io/endpoint?url=https://raw.githubusercontent.com/mmogr/gglib/badges/gglib-core-domain-agent-stagnation-complexity.json)
 
 <!-- module-docs:start -->
 
@@ -10,7 +10,7 @@ Text stagnation detection for the agentic loop.
 # Algorithm
 
 After each LLM response the assistant's text is hashed with
-[`crate::fnv1a::fnv1a_64`].  The hash is looked up in a session-wide
+[`crate::domain::agent::fnv1a::fnv1a_64`].  The hash is looked up in a session-wide
 occurrence map.  When a hash has already been seen at least
 `max_stagnation_steps` times **before** the current call, the loop is
 aborted with [`AgentError::StagnationDetected`].
@@ -18,7 +18,7 @@ aborted with [`AgentError::StagnationDetected`].
 Stagnation detection runs on **every** iteration — both tool-calling and
 text-only — so it catches models that repeat the same text regardless of
 whether tools are invoked.  Tool-call loops are handled separately by
-[`crate::loop_detection::LoopDetector`].
+[`crate::domain::agent::loop_detection::LoopDetector`].
 
 ## Oscillation detection
 
@@ -41,7 +41,7 @@ iterations do not accumulate spurious counts.
 <!-- module-table:start -->
 | Module | LOC | Complexity | Coverage |
 |--------|-----|------------|----------|
-| [`tests.rs`](tests.rs) | ![](https://img.shields.io/endpoint?url=https://raw.githubusercontent.com/mmogr/gglib/badges/gglib-agent-stagnation-tests-loc.json) | ![](https://img.shields.io/endpoint?url=https://raw.githubusercontent.com/mmogr/gglib/badges/gglib-agent-stagnation-tests-complexity.json) | ![](https://img.shields.io/endpoint?url=https://raw.githubusercontent.com/mmogr/gglib/badges/gglib-agent-stagnation-tests-coverage.json) |
+| [`tests.rs`](tests.rs) | ![](https://img.shields.io/endpoint?url=https://raw.githubusercontent.com/mmogr/gglib/badges/gglib-core-stagnation-tests-loc.json) | ![](https://img.shields.io/endpoint?url=https://raw.githubusercontent.com/mmogr/gglib/badges/gglib-core-stagnation-tests-complexity.json) | ![](https://img.shields.io/endpoint?url=https://raw.githubusercontent.com/mmogr/gglib/badges/gglib-core-stagnation-tests-coverage.json) |
 <!-- module-table:end -->
 
 </details>
