+++ crates/gglib-core/src/domain/agent/stagnation/tests.rs
