+++ crates/gglib-core/src/domain/mod.rs
@@ -113,9 +113,10 @@ pub use model_naming::{
 pub use agent::{
     AGENT_EVENT_CHANNEL_CAPACITY, AgentConfig, AgentConfigError, AgentEvent, AgentMessage,
     AssistantContent, DEFAULT_MAX_ITERATIONS, DEFAULT_MAX_PARALLEL_TOOLS,
-    DEFAULT_MAX_STAGNATION_STEPS, LlmStreamEvent, MAX_ITERATIONS_CEILING,
+    DEFAULT_MAX_STAGNATION_STEPS, LlmStreamEvent, LoopDetector, MAX_ITERATIONS_CEILING,
     MAX_PARALLEL_TOOLS_CEILING, MAX_TOOL_TIMEOUT_MS_CEILING, MIN_CONTEXT_BUDGET_CHARS,
-    MIN_TOOL_TIMEOUT_MS, ToolCall, ToolDefinition, ToolResult,
+    MIN_TOOL_TIMEOUT_MS, StagnationDetector, ToolCall, ToolDefinition, ToolResult,
+    is_observation_batch,
 };
 
 // Re-export capability types at the domain level for convenience
